refactor: move common logic to typeUtils
